refactor: policies tab content
