+++ src/components/pages/earn/ComputeResourceNodeDetailPage/cmp.tsx
@@ -1,6 +1,6 @@
 import { memo } from 'react'
 import Head from 'next/head'
-import { useComputeResourceNodeDetailPage } from '@/hooks/pages/earn/useComputeResourceNodeDetailPage'
+import { useComputeResourceNodeDetailPage } from '@/hooks/pages/earn/ComputeResourceNodeDetailPage/useComputeResourceNodeDetailPage'
 import { Button, Icon, Tabs } from '@aleph-front/core'
 import NodeDetailHeader from '@/components/common/NodeDetailHeader'
 import OverviewTabContent from './tabs/OverviewTabContent'
